src/main/java/com/mysema/dropbook/security/DropbookUserDeserializer.java
package com.mysema.dropbook.security;

import com.fasterxml.jackson.core.JsonParser;
import com.fasterxml.jackson.core.JsonProcessingException;
import com.fasterxml.jackson.core.ObjectCodec;
import com.fasterxml.jackson.databind.DeserializationContext;
import com.fasterxml.jackson.databind.JsonDeserializer;
import com.fasterxml.jackson.databind.JsonNode;

import java.io.IOException;
import java.util.Collections;

public class DropbookUserDeserializer extends JsonDeserializer<DropbookUser> {
    @Override
    public DropbookUser deserialize(JsonParser jsonParser, DeserializationContext deserializationContext) throws IOException, JsonProcessingException {
        ObjectCodec codec = jsonParser.getCodec();
        JsonNode node = codec.readTree(jsonParser);
        return new DropbookUser(null, node.get("username").textValue(), node.get("password").textValue(), null);
    }
}
